sensor/pyproject.toml
[project]
name = "sensor"
version = "0.1.0"
description = "Sensor template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "datetime>=5.5",
    "fastapi>=0.115.6",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
